+++ packages/frontend-shared/src/locales/en-US.json
@@ -222,6 +222,7 @@
       "dropText": "Drag your project directory here or {0}",
       "browseManually": "browse manually."
     },
+    "selectPreferredEditor": "Select Preferred Editor",
     "recentProjectsHeader": "Recent Projects",
     "searchPlaceholder": "Search projects",
     "addProjectButton": "Add Project",
